✨ Feat: question components의 class를 abstract class로 만들어 적용
